# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyWave Q&A Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  username: string;
  email: string;
  password: string;
  displayName: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  points: number;
  role: string;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Badge {
  id: number;
  name: string;
  description: string;
  icon: string;
  color: string;
  pointsRequired: number;
  category: string;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  points: number;
  role: string;
  createdAt: string;
  questionCount: number;
  answerCount: number;
  awardedAnswerCount: number;
  badges: Badge[];
}

export interface UpdateSettingsBody {
  displayName?: string;
  bio?: string;
  email?: string;
  currentPassword?: string;
  newPassword?: string;
}

export interface UploadAvatarBody {
  avatarUrl: string;
}

export interface Question {
  id: number;
  title: string;
  content: string;
  subject: string;
  authorId: number;
  authorUsername: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  upvotes: number;
  downvotes: number;
  answerCount: number;
  hasAwardedAnswer: boolean;
  isSolved: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface Answer {
  id: number;
  content: string;
  questionId: number;
  authorId: number;
  authorUsername: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  authorPoints: number;
  upvotes: number;
  downvotes: number;
  isAwarded: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface QuestionDetail {
  id: number;
  title: string;
  content: string;
  subject: string;
  authorId: number;
  authorUsername: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  authorPoints: number;
  upvotes: number;
  downvotes: number;
  hasAwardedAnswer: boolean;
  isSolved: boolean;
  createdAt: string;
  updatedAt: string;
  answers: Answer[];
}

export interface PaginatedQuestions {
  questions: Question[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateQuestionBody {
  title: string;
  content: string;
  subject: string;
}

export interface UpdateQuestionBody {
  title?: string;
  content?: string;
  subject?: string;
}

export interface VoteBody {
  type: string;
}

export interface CreateAnswerBody {
  questionId: number;
  content: string;
}

export interface UpdateAnswerBody {
  content?: string;
}

export interface CreateBadgeBody {
  name: string;
  description: string;
  icon: string;
  color: string;
  pointsRequired: number;
  category: string;
}

export interface UpdateBadgeBody {
  name?: string;
  description?: string;
  icon?: string;
  color?: string;
  pointsRequired?: number;
  category?: string;
}

export interface PlatformStats {
  totalUsers: number;
  totalQuestions: number;
  totalAnswers: number;
  totalAwardedAnswers: number;
  solvedQuestions: number;
  activeUsersToday: number;
}

export interface LeaderboardEntry {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  points: number;
  answerCount: number;
  awardedAnswerCount: number;
  rank: number;
}

export interface SubjectStat {
  subject: string;
  count: number;
  solvedCount: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  userId: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  questionId?: number | null;
  /** @nullable */
  questionTitle?: string | null;
  createdAt: string;
}

export interface PaginatedUsers {
  users: User[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AdminUpdateUserBody {
  displayName?: string;
  email?: string;
  role?: string;
  isActive?: boolean;
  points?: number;
}

export interface AdminStats {
  totalUsers: number;
  totalQuestions: number;
  totalAnswers: number;
  totalAwardedAnswers: number;
  solvedQuestions: number;
  activeUsersToday: number;
  newUsersThisWeek: number;
  newQuestionsThisWeek: number;
  newAnswersThisWeek: number;
  totalBadges: number;
  topSubjects: SubjectStat[];
}

export type ListQuestionsParams = {
  subject?: string;
  search?: string;
  page?: number;
  limit?: number;
  sort?: string;
};

export type AdminListUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type AdminListQuestionsParams = {
  page?: number;
  limit?: number;
};