# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyWave Q&A Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin - list all users
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin - update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin - delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/questions:
    get:
      operationId: adminListQuestions
      tags: [admin]
      summary: Admin - list all questions
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedQuestions"
  /admin/questions/{id}:
    delete:
      operationId: adminDeleteQuestion
      tags: [admin]
      summary: Admin - delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/badges:
    post:
      operationId: adminCreateBadge
      tags: [admin]
      summary: Admin - create a badge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBadgeBody"
      responses:
        "201":
          description: Created badge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Badge"
  /admin/badges/{id}:
    patch:
      operationId: adminUpdateBadge
      tags: [admin]
      summary: Admin - update a badge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBadgeBody"
      responses:
        "200":
          description: Updated badge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Badge"
    delete:
      operationId: adminDeleteBadge
      tags: [admin]
      summary: Admin - delete a badge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        points:
          type: integer
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - displayName
        - points
        - role
        - isActive
        - createdAt
...
    Question:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        subject:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorDisplayName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        upvotes:
          type: integer
        downvotes:
          type: integer
        answerCount:
          type: integer
        hasAwardedAnswer:
          type: boolean
        isSolved:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - subject
        - authorId
        - authorUsername
        - authorDisplayName
        - upvotes
        - downvotes
        - answerCount
        - hasAwardedAnswer
        - isSolved
        - createdAt
        - updatedAt
...
    PaginatedQuestions:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - questions
        - total
        - page
        - limit
        - totalPages
...
    Badge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        pointsRequired:
          type: integer
        category:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - icon
        - color
        - pointsRequired
        - category
        - createdAt
    CreateBadgeBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        pointsRequired:
          type: integer
        category:
          type: string
      required:
        - name
        - description
        - icon
        - color
        - pointsRequired
        - category
    UpdateBadgeBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        pointsRequired:
          type: integer
        category:
          type: string
...
    PaginatedUsers:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - users
        - total
        - page
        - limit
        - totalPages
    AdminUpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        points:
          type: integer
...