# lib/api-zod/src/generated/types/createBadgeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyWave Q&A Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBadgeBody {
  name: string;
  description: string;
  icon: string;
  color: string;
  pointsRequired: number;
  category: string;
}