# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudyWave Q&A Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: number;
  title: string;
  content: string;
  subject: string;
  authorId: number;
  authorUsername: string;
  authorDisplayName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  upvotes: number;
  downvotes: number;
  answerCount: number;
  hasAwardedAnswer: boolean;
  isSolved: boolean;
  createdAt: string;
  updatedAt: string;
}